Made python files vectors to fit poblem dimension. TO-DO: 2D case and 3D case always use 3 component vectos instead of 2D/3D

+++ kratos/processes/structured_mesh_generator_process.h
@@ -126,7 +126,7 @@ namespace Kratos
       ///@name Member Variables
       ///@{
       const GeometryType& mrGeometry;
-      std::vector<int> mNumberOfDivisions;
+      std::vector<size_t> mNumberOfDivisions;
 		  std::size_t mStartNodeId;
 		  std::size_t mStartElementId;
 		  std::size_t mStartConditionId;
